Corrects usage of retryAmount
"retryAmount" implies that there is some number of retries that should be attempted.  With the prefix decrement, this is not the case.  That operator treats "retryAmount" as the total number of query attempts.  I have switched it to a postfix decrement to ensure that it instead reflects the number of retries that should be performed.
A minor typo in log message was addressed, too.